+++ src/parser/tests/parser_tests.rs
@@ -3576,7 +3576,7 @@ match x, y
   0, 1 or 2, 3 if z then 0
   a, ()
     a
-  _ then 0
+  else 0
 ";
             check_ast(
                 source,
@@ -3596,8 +3596,7 @@ match x, y
                     Empty,
                     TempTuple(vec![11, 12]),
                     Id(5),
-                    Wildcard, // 15
-                    Number0,
+                    Number0, // 15
                     Match {
                         expression: 2,
                         arms: vec![
@@ -3612,14 +3611,14 @@ match x, y
                                 expression: 14,
                             },
                             MatchArm {
-                                patterns: vec![15],
+                                patterns: vec![],
                                 condition: None,
-                                expression: 16,
+                                expression: 15,
                             },
                         ],
                     },
                     MainBlock {
-                        body: vec![17],
+                        body: vec![16],
                         local_count: 1,
                     },
                 ],
@@ -3639,7 +3638,7 @@ match x, y
             let source = "
 match x.foo 42
   () then 0
-  _ then 1
+  else 1
 ";
             check_ast(
                 source,
@@ -3651,7 +3650,6 @@ match x.foo 42
                     Lookup((LookupNode::Root(0), Some(3))),
                     Empty, // 5
                     Number0,
-                    Wildcard,
                     Number1,
                     Match {
                         expression: 4,
@@ -3662,14 +3660,14 @@ match x.foo 42
                                 expression: 6,
                             },
                             MatchArm {
-                                patterns: vec![7],
+                                patterns: vec![],
                                 condition: None,
-                                expression: 8,
+                                expression: 7,
                             },
                         ],
                     },
                     MainBlock {
-                        body: vec![9],
+                        body: vec![8],
                         local_count: 0,
                     },
                 ],